MDL-55616 assign: Template directory for batch downloaded assignments.
This allows you to edit a template to set the folder name that will be in
a batch of downloaded assignments.

A suffix is appended to the template on processing that contains the
submission uniqueid for user, plugin subtype and plugin type.
An example of this string is _293002_assignsubmission_onlinetext_

The suffix is used to determine who the submission belongs to and what
plugin type to use when importing zip feedback.

The following variables can be set:
{{username}} The submisisons username.
{{idnumber}} The submissions users idnumber.
{{groupname}} The group the submission belongs to.
{{course}} The course shortname.
{{userid}} The submissions userid.
{{assignmentname}} The name of the assignment.
{{attemptnumber}} The number of submitted attempts.
{{fullname}} The full name of the user.
{{profileXXXX}} Where XXXX is the shortname of a custom user profile field.
{{submissioncreated}} The date of the initial submission.
{{submissionmodified}} The date of the submission after a modification.

If the {{profileXXXX}} fields are not found then the token is removed
from  the template string when it processes the folder name.

{{username}} and {{idnumber}} respect $CFG->showuseridentity.