MDL-55616 assign: Template directory for batch downloaded assignments.
This allows you to edit a template to set the folder name that will be in
a batch of downloaded assignments.

A suffix is appended to the template on processing that contains the
submission uniqueid for user, plugin subtype and plugin type.
An example of this string is _293002_assignsubmission_onlinetext_

The suffix is used to determine who the submission belongs to and what
plugin type to use when importing zip feedback.

The following template items exist for each submission folder in a downloaded archive:
{{username}} The username of the user.
{{idnumber}} The idnumber of the user.
{{userid}} The internal id number of the user.
{{fullname}} The full name of the user.
{{course}} The course shortname.
{{groupname}} The group name that the submitting user belongs to.
{{submissioncreated}} The date of the initial submission.
{{submissionmodified}} The date of the submission after a modification.
{{assignmentname}} The name of the assignment.
{{attemptnumber}} The number of submitted attempts.
{{profileXXXX}} Where XXXX is the shortname of a user custom profile field.

If the {{profileXXXX}} fields are not found then the token is removed
from  the template string when it processes the folder name.

{{username}} and {{idnumber}} respect $CFG->showuseridentity.